Removing the podspec. Read below for why.
Because HEAnalytics is written in Swift, cocoapods requires the use of frameworks. All of the analytics SDKs offer themselves as static libraries. And cocoapods doesn't jive with this due to dependency and linker resolution issues. Who knows... there may be a way to deal with this, but after much digging around I cannot figure out a solution other than waiting for all the vendors to release REAL frameworks (not fake frameworks that just embed .a files) -- but who knows when that'll happen.

I wonder if Carthage will be any better. :-)